var range fix
git-svn-id: svn://svn.code.sf.net/p/cdcat/svn/trunk@214 23e97676-e3e9-41ff-bed1-4a4288a26c68

+++ src/dbase.cpp
@@ -846,7 +846,7 @@ int DataBase::scanFsToNode ( QString what,Node *to ) {
             if (*DEBUG_INFO_ENABLED)
 		std::cerr << "adding file: " << qPrintable ( fileInfo->fileName() ) << std::endl;
             
-            float size = fileInfo->size();
+            long size = fileInfo->size();
 
             if (*DEBUG_INFO_ENABLED)
 		std::cerr << "adding file size: " << fileInfo->size() << std::endl;
